refactor: extract app domain state hook
